toollama/moe/requirements.txt
# Web Framework
flask>=2.0.0
flask-cors>=3.0.0

# HTTP Client
httpx>=0.24.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=0.19.0
loguru>=0.7.0

# Model Dependencies
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0

# Type Checking
mypy>=1.0.0
types-PyYAML>=6.0.0
types-requests>=2.31.0

# Development Tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0 